+++ src/dummy-fusebox/fusebox.jsx
@@ -5,8 +5,22 @@ import * as api from '../wireflow';
 
 const {
   Dispatcher,
-  model: { convert, Settings, READ_ONLY, READ_WRITE },
-  stores: { LayoutStore, GraphStore, SelectionStore, HistoryStore },
+  actions: {
+    CreateCheckpoint,
+    AutoLayout
+  },
+  model: {
+    convert,
+    Settings,
+    READ_ONLY,
+    READ_WRITE
+  },
+  stores: {
+    LayoutStore,
+    GraphStore,
+    SelectionStore,
+    HistoryStore
+  },
   components: { Graph }
 } = api;
 
@@ -15,22 +29,28 @@ const {
 const graph = convert.graph( data.graph );
 const layout = convert.layout( data.layout );
 const types = convert.types( data.types );
+var settings = Settings({ mode: READ_ONLY });
 
 const dispatcher = new Dispatcher( render );
 
 new HistoryStore( dispatcher );
 const graphStore = new GraphStore( dispatcher, graph, types );
-const layoutStore = new LayoutStore( dispatcher, layout, types );
+const layoutStore = new LayoutStore( dispatcher, layout, graphStore );
 const selectionStore = new SelectionStore( dispatcher, layoutStore, graphStore );
 
-
-var settings = Settings({ mode: READ_ONLY });
-const toggleMode = () => {
+function toggleMode() {
   settings = Settings({
     mode: settings.mode === READ_ONLY ? READ_WRITE : READ_ONLY
   });
   render();
-};
+}
+
+function autoLayout() {
+  dispatcher.dispatch( CreateCheckpoint() );
+  dispatcher.dispatch( AutoLayout() );
+  render();
+}
+
 
 function render() {
   // Later: <History checkpoints={historyStore.checkpoints } now={ historyStore.now } />
@@ -38,6 +58,7 @@ function render() {
   React.render(
     <div className='demo-wrapper'>
       <div className='demo-menu'>
+        <button onClick={autoLayout}>Auto-Layout</button>
         <label>
           <input type="checkbox"
                  checked={settings.mode === READ_ONLY}