+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Last Changes
 
+- implemented automatic graph layout (through dagre)
 - worked around update/measurement-related bug
 - improved vertex layout by using display: table